Always follow the core's choice regarding CUDA/OpenMP support
Allowing the library to disable features that the core was compiled
with proved harder than expected, since `legate_defines.h` overrides the
relevant `LEGATE_USE_*` flags according to what Legion was compiled with.

It is also possible that setting a different value than what the core
was compiled with will lead to the library and the core using different
layouts for the same struct, leading to weird errors.