src/cunumeric: handle high number of bins in GPU bincount
The existing bincount implementation on GPUs attempts to allocate
a workspace for all bins within the shared memory available on each
SM. This commit updates the implementation to fall back to a slower
kernel that reduces to global memory when there are too many bins to
fit into shared memory.

Fixes #503.