Handle the case where LAPACK_*potrf is a macro, not a function (#527) (#539)
This is working around some recent changes to OpenBLAS. Previously we were using
the internal names for functions, e.g. "spotrf_". OpenBLAS changed the
definitions of these internal functions, so in a previous PR we switched to
using the public functions, e.g. "LAPACK_spotrf". These used to be function
symbols, but in the latest update OpenBLAS changed these to be macros.

Co-authored-by: Manolis Papadakis <manopapad@gmail.com>