Invoke eye with read-write privilege, not write-discard
We cannot create tight region requirements, that include just the diagonal we
are writing, so necessarily there will be elements in the regions we pass to the
eye call whose values must remain. Write-discard privilege, then, is not
appropriate for this call, as it essentially tells the runtime that it can throw
away the previous contents of the entire region.