Reject types that DeferredArray doesn't support
IMHO there is little value in pretending that we support these types.
We currently accept them by falling back to EagerArray, and that
works for a little bit, but it is possible that later we will need to
convert to a truly Legion-backed implementation, in which case the
failure will pop up there.

Instead we should just reject them from the start.